refactor: move array helpers
